+++ src/main/java/lavaplayer/PlayerManager.java
@@ -1,0 +1,100 @@
+package lavaplayer;
+
+import com.sedmelluq.discord.lavaplayer.player.AudioLoadResultHandler;
+import com.sedmelluq.discord.lavaplayer.player.AudioPlayerManager;
+import com.sedmelluq.discord.lavaplayer.player.DefaultAudioPlayerManager;
+import com.sedmelluq.discord.lavaplayer.source.AudioSourceManagers;
+import com.sedmelluq.discord.lavaplayer.tools.FriendlyException;
+import com.sedmelluq.discord.lavaplayer.track.AudioPlaylist;
+import com.sedmelluq.discord.lavaplayer.track.AudioTrack;
+import net.dv8tion.jda.api.entities.Guild;
+import net.dv8tion.jda.api.entities.TextChannel;
+
+import java.awt.dnd.InvalidDnDOperationException;
+import java.util.HashMap;
+import java.util.Map;
+
+public class PlayerManager {
+    private static PlayerManager INSTANCE;
+    private final AudioPlayerManager playerManager;
+    private final Map<Long, GuildMusicManager> musicManagers;
+
+    private PlayerManager() {
+        this.musicManagers = new HashMap<>();
+
+        this.playerManager = new DefaultAudioPlayerManager();
+        AudioSourceManagers.registerRemoteSources(playerManager);
+        AudioSourceManagers.registerLocalSource(playerManager);
+    }
+
+    public static synchronized PlayerManager getInstance() {
+        if (INSTANCE == null) {
+            INSTANCE = new PlayerManager();
+        }
+
+        return INSTANCE;
+    }
+
+    public synchronized GuildMusicManager getGuildMusicManager(Guild guild) {
+        long guildId = guild.getIdLong();
+        GuildMusicManager musicManager = musicManagers.get(guildId);
+
+        if (musicManager == null) {
+            musicManager = new GuildMusicManager(playerManager);
+            musicManagers.put(guildId, musicManager);
+        }
+
+        guild.getAudioManager().setSendingHandler(musicManager.getSendHandler());
+
+        return musicManager;
+    }
+
+    public void loadAndPlay(final TextChannel channel, final String trackUrl) {
+        GuildMusicManager f = getGuildMusicManager(channel.getGuild());
+        final GuildMusicManager musicManager = f;
+        //DefaultAudioPlayerManager maanager = new Youtube();
+        //playerManager.registerSourceManager(maanager);
+        playerManager.loadItemOrdered(musicManager, trackUrl, new AudioLoadResultHandler() {
+            @Override
+            public void trackLoaded(AudioTrack track) {
+                channel.sendMessage(":thumbsup: Tocando agora: " + track.getInfo().title).queue();
+
+                play(musicManager, track);
+            }
+
+            @Override
+            public void playlistLoaded(AudioPlaylist playlist) {
+                AudioTrack firstTrack = playlist.getSelectedTrack();
+
+                if (firstTrack == null) {
+                    firstTrack = playlist.getTracks().get(0);
+                }
+
+                channel.sendMessage("Ok vou colocar \"" + firstTrack.getInfo().title + "\" pra tocar, e depois o resto da playlist \"" + playlist.getName() + "\", beleza?").queue();
+
+                play(musicManager, firstTrack);
+                //pra tocar as próximas músicas, absolutamente não sei pq isso não tava aqui já
+                for (int g = 1; g < playlist.getTracks().size(); g++) {
+                    musicManager.scheduler.queue(playlist.getTracks().get(g));
+                }
+
+            }
+
+            @Override
+            public void noMatches() {
+                //channel.sendMessage("Não achei nada com esse link, cê digitou certo? Olha: " + trackUrl).queue();
+                throw new InvalidDnDOperationException();
+            }
+
+            @Override
+            public void loadFailed(FriendlyException exception) {
+                channel.sendMessage("Deu pra tocar não mano :(, " + exception.getMessage()).queue();
+            }
+        });
+
+    }
+
+    private void play(GuildMusicManager musicManager, AudioTrack track) {
+        musicManager.scheduler.queue(track);
+    }
+}
